added action bar menu for logout

+++ app/src/main/java/com/example/notices/NoticeList.java
@@ -3,13 +3,14 @@ package com.example.notices;
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 
 public class NoticeList extends AppCompatActivity {
 
     private SQLiteHandler db;
-    private Button logout;
     private SessionManager session;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -18,21 +19,36 @@ public class NoticeList extends AppCompatActivity {
 
         db = new SQLiteHandler(getApplicationContext());
         session = new SessionManager(getApplicationContext());
-        logout = (Button) findViewById(R.id.logout);
-        logout.setOnClickListener(
-                new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        session.setLogin(false);
 
-                        db.deleteUsers();
-
-                        // Launching the login activity
-                        Intent intent = new Intent(getApplicationContext(), login.class);
-                        startActivity(intent);
-                        finish();
-                    }
-                }
-        );
     }
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.actionmenu, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        int id = item.getItemId();
+
+        //noinspection SimplifiableIfStatement
+        if (id == R.id.logout) {
+
+            session.setLogin(false);
+
+            db.deleteUsers();
+
+            // Launching the login activity
+            Intent intent = new Intent(getApplicationContext(), login.class);
+            startActivity(intent);
+            finish();
+        }
+
+        return true;
+    }
+
 }
